requirements.txt
streamlit
numpy
nltk
faiss-cpu
PyMuPDF
reportlab
langchain
langchain-groq
sentence-transformers
python-dotenv
sendgrid
transformers
torch